[SQUASH] Implement click to take partial screenshot [1/2]
Default to the old long press behaviour when the feature is enabled.

Change-Id: I5fc37f818b06db2bd2d315cffee2041a740781ed
Signed-off-by: Jesse Chan <jc@lineageos.org>
Signed-off-by: Hưng Phan <phandinhhungvp2001@gmail.com>

SystemUI: Dismiss existing screenshot window before new

To make sure the windows won't appear in next screenshot.

Change-Id: Ic065816cb1c808540a82383a7f24f1bce0dd6589
Signed-off-by: LibXZR <i@xzr.moe>
Signed-off-by: Adithya R <gh0strider.2k18.reborn@gmail.com>
Signed-off-by: Hưng Phan <phandinhhungvp2001@gmail.com>

SystemUI: screenshot: open the screenshot instead of edit

Open the taken screenshot in gallery application when clicking on the
preview. There is a seperate edit button, so the preview imageview
should open gallery instead.

Test: atest com.android.systemui.screenshot.ScreenshotNotificationSmartActionsTest
Change-Id: I8f882ddd0da47ca50acc37d25ee0866ce5698e4f
Signed-off-by: Hưng Phan <phandinhhungvp2001@gmail.com>

SystemUI: screenshot: open long screenshot activity for partial screenshots

The partial screenshot UI is basically unusable. Replace it
with the long  screenshot UI which offers cropping, saving,
canceling, and sharing capabilities.

[Stallix] Adapt for A13 QPR3

Change-Id: I27f7a811838f7e3f2057f567c7a350255996a8cb
Signed-off-by: Hưng Phan <phandinhhungvp2001@gmail.com>

SystemUI: screenshot: remove duplicate clamping on move action

On MotionEvent.ACTION_DOWN, a range of of motion is calculated and
stored. On MotionEvent.ACTION_MOVE, the new position of the crop
boundary is clamped by the original range of motion, and also
clamped again by a new range of motion inside setBoundaryPosition().

The original range of motion is pointless and only used in this case,
remove it.

Change-Id: Ib4f35bf5317f5b5879606680beee3e077f829d10
Signed-off-by: Hưng Phan <phandinhhungvp2001@gmail.com>

SystemUI: screenshot: add extra crop boundaries

Add corner boundaries that allow resizing in both vertical and
horizontal directions at the same time.

To handle corner boundaries, split every boundary into vertical
and horizontal boundaries, and handle both vertical and horizontal
changes for all boundaries. If the boundary is a purely vertical
(top, bottom) or purely horizontal (left, right) boundary, the
opposite direction boundary will be none, and the selection won't
be changed on that selection.

Also, add a middle boundary that allows moving the current selection.

To achieve this, treat the middle boundary as composed of top and
left boundaries, and when trying to change either the top or the
left boundary while the currently dragged boundary is the middle
one, reposition the opposite boundary too.

Change-Id: I6dd703733b3d161f5979d91a124350fb1a8d52ae
Signed-off-by: Hưng Phan <phandinhhungvp2001@gmail.com>

SystemUI: screenshot: disable magnification for partial screenshots

Partial screenshots are screen-size and don't need a magnifying glass.

Change-Id: I472103352411d1482423230a4ff08ddda45cc53d
Signed-off-by: Hưng Phan <phandinhhungvp2001@gmail.com>